Stabilized engine and optimized rendering; got ai following.
Crashes are a thing of the past! Cached all the textures as quads in the Textures table, precomputed the rays casted out for less cpu usage and to address prior memory leaks, and the red square follows propely. Will fine tune it once I get sprites and multiple enemies spawned and stable.